Continue's button now does not display onto the game. Continue's button in the character selection can't be clicked unless a card is chosen. Abandon now well reloads back to the house selection page.
